% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{checkInputs}
\alias{checkInputs}
\title{Check input quality}
\usage{
checkInputs(x, ...)
}
\arguments{
\item{x}{object to check}

\item{...}{additional parameters}
}
\description{
Check input quality
}

