% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{checkInputs_file_mode}
\alias{checkInputs_file_mode}
\title{Check inputs for file mode}
\usage{
checkInputs_file_mode(x, ...)
}
\arguments{
\item{x}{object to check}

\item{...}{additional parameters}
}
\description{
Check inputs for file mode
}

