% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{checkInputs_manual_mode}
\alias{checkInputs_manual_mode}
\title{Check inputs for manual mode}
\usage{
checkInputs_manual_mode(x, ...)
}
\arguments{
\item{x}{object to check}

\item{...}{additional parameters}
}
\description{
Check inputs for manual mode
}

