% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{checkInputs_matrix_mode}
\alias{checkInputs_matrix_mode}
\title{Check inputs for matrix mode}
\usage{
checkInputs_matrix_mode(x, ...)
}
\arguments{
\item{x}{object to check}

\item{...}{additional parameters}
}
\description{
Check inputs for matrix mode
}

