% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructModel.R
\name{constructModel}
\alias{constructModel}
\title{Generic function.}
\usage{
constructModel(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{other arguments}
}
\value{
Returns an object of class OP that is supported by the ROI class
}
\description{
Generic function.
}

