% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{getProblemSummary}
\alias{getProblemSummary}
\title{Get problem summary}
\usage{
getProblemSummary(out)
}
\arguments{
\item{out}{object returned by AlteryxSolve}
}
\description{
Get problem summary
}

