% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{getProblemType}
\alias{getProblemType}
\title{Detect problem type based on class of OP object and types of variables}
\usage{
getProblemType(out)
}
\arguments{
\item{out}{object returned by AlteryxSolve}
}
\description{
Detect problem type based on class of OP object and types of variables
}

