% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDataNew.R
\name{inferA}
\alias{inferA}
\title{Infer column names for Input A based on constraintMode}
\usage{
inferA(A, constrMode)
}
\arguments{
\item{A}{A data.frame, original A matrix from input Anchor A}

\item{constrMode}{A string, constraint mode.}
}
\description{
Infer column names for Input A based on constraintMode
}

