% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDataNew.R
\name{inferB}
\alias{inferB}
\title{Infer column names for Input B to 'constraint', 'rhs', 'dir'}
\usage{
inferB(B)
}
\arguments{
\item{B}{A data.frame, original B matrix from input Anchor B}
}
\description{
Infer column names for Input B to 'constraint', 'rhs', 'dir'
}
\examples{
\dontrun{
 B <- data.frame(
  z = c('A', 'B', 'C'),
  x = c(">=", "<=", "=="),
  y = c(1, 2, 3)
 )
 inferB(B)
}
}

