% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDataNew.R
\name{inferO}
\alias{inferO}
\title{Infer column names for Input O to "variable", "coefficient", "lb", "ub", "type"}
\usage{
inferO(O, config)
}
\arguments{
\item{O}{A data.frame, original O matrix from input Anchor O.}

\item{config}{configuration object as a list.}
}
\description{
Infer column names for Input O to "variable", "coefficient", "lb", "ub", "type"
}

