% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualInputs.R
\name{makeCPLEXFile}
\alias{makeCPLEXFile}
\title{Transform model into CPLEX LP format}
\usage{
makeCPLEXFile(x)
}
\arguments{
\item{x}{list of objective, constraints and fieldInfo}
}
\description{
Transform model into CPLEX LP format
}

