% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{makeDataOutput}
\alias{makeDataOutput}
\title{Make data output}
\usage{
makeDataOutput(out, asJSON = FALSE)
}
\arguments{
\item{out}{object returned by AlteryxSolve}
}
\description{
Make data output
}

