% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{makeInteractiveReport}
\alias{makeInteractiveReport}
\title{Make interactive report}
\usage{
makeInteractiveReport(out)
}
\arguments{
\item{out}{output from AlteryxSolve}
}
\description{
Make interactive report
}

