% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{makeVariableReport}
\alias{makeVariableReport}
\title{Make variable report}
\usage{
makeVariableReport(out)
}
\arguments{
\item{out}{output from AlteryxSolve}
}
\description{
Make variable report
}

