% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveModel.R
\name{solveModel}
\alias{solveModel}
\title{Solve model}
\usage{
solveModel(x, solver, ...)
}
\arguments{
\item{x}{x}

\item{solver}{solver}

\item{...}{additional arguments. currently not used.}
}
\description{
Solve model
}

