% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive-report.R
\name{makeInteractiveReport}
\alias{makeInteractiveReport}
\title{Make interactive report}
\usage{
makeInteractiveReport(out, nOutput = 3, ...)
}
\arguments{
\item{out}{output from AlteryxSolve}

\item{nOutput}{output anchor to send html output to}

\item{...}{other arguments to pass to renderInComposer}
}
\description{
Make interactive report
}

